requirements.txt
mysql-connector-python

supabase
